+++ packages/expo-router/CHANGELOG.md
@@ -38,6 +38,7 @@
 - allow for standalone usage of Stack.Header and Stack.Header.Left/Right ([#41649](https://github.com/expo/expo/pull/41649) by [@Ubax](https://github.com/Ubax))
 - Unify the API between Toolbar, Header items and Link preview ([#41679](https://github.com/expo/expo/pull/41679) by [@Ubax](https://github.com/Ubax))
 - [android] Add Material 3 dynamic color defaults for NativeTabs ([#41869](https://github.com/expo/expo/pull/41869) by [@Ubax](https://github.com/Ubax))
+- [iOS] add Toolbar.SearchBar ([#41863](https://github.com/expo/expo/pull/41863) by [@Ubax](https://github.com/Ubax))
 
 ### 🐛 Bug fixes